fix: improve security with mistake with endpoints insecure in News
Prevent attacks Broken Object Level Authorization in endpoints for news

+++ backend/src/news/news.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateNewsDto } from './dto/create-news.dto';
 import { UpdateNewsDto } from './dto/update-news.dto';
 import { PrismaService } from 'prisma/prisma.service';
@@ -7,9 +7,12 @@ import { PrismaService } from 'prisma/prisma.service';
 export class NewsService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async create(createNewsDto: CreateNewsDto) {
+  async create(createNewsDto: CreateNewsDto, authorId: string) {
     return await this.prisma.news.create({
-      data: createNewsDto,
+      data: {
+        ...createNewsDto,
+        authorId,
+      },
     });
   }
 
@@ -34,14 +37,46 @@ export class NewsService {
     });
   }
 
-  async update(id: string, updateNewsDto: UpdateNewsDto) {
+  async update(
+    id: string,
+    updateNewsDto: UpdateNewsDto,
+    userId: string,
+    userRole: string,
+  ) {
+    const news = await this.prisma.news.findUnique({
+      where: { id },
+    });
+
+    if (!news) {
+      throw new HttpException('News not found', HttpStatus.NOT_FOUND);
+    }
+
+    if (news.authorId !== userId && userRole !== 'admin') {
+      throw new HttpException(
+        'You are not allowed to update this news',
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
     return await this.prisma.news.update({
       where: { id },
       data: updateNewsDto,
     });
   }
 
-  async remove(id: string) {
+  async remove(id: string, userId: string, userRole: string) {
+    const news = await this.prisma.news.findUnique({where: { id }});
+    
+    if (!news) {
+      throw new HttpException('News not found', HttpStatus.NOT_FOUND);
+    }
+
+    if (news.authorId !== userId && userRole !== 'admin') {
+      throw new HttpException(
+        'You are not allowed to delete this news',
+        HttpStatus.FORBIDDEN,
+      );
+    }
     return await this.prisma.news.delete({
       where: { id },
     });
